feat: move metrics out of their sections
