+++ pages/dataservices/[did].vue
@@ -202,6 +202,16 @@
                   </dd>
                 </template>
               </div>
+              <div>
+                <StatBox
+                  :title="$t('Views')"
+                  :data="metricsViews"
+                  size="sm"
+                  type="line"
+                  :summary="metricsViewsTotal"
+                  class="mb-8 md:mb-0"
+                />
+              </div>
             </dl>
           </div>
         </div>
@@ -273,7 +283,7 @@
 </template>
 
 <script setup lang="ts">
-import { isOrganizationCertified, BrandedButton, Swagger, ReadMore, SimpleBanner, type Dataservice, AvatarWithName, useFormatDate, type DataserviceAccessAudienceCondition, type DataserviceAccessAudienceType, type DataserviceAccessAudience } from '@datagouv/components-next'
+import { isOrganizationCertified, BrandedButton, Swagger, ReadMore, SimpleBanner, type Dataservice, AvatarWithName, useFormatDate, type DataserviceAccessAudienceType, type DataserviceAccessAudience, StatBox } from '@datagouv/components-next'
 import { RiArrowDownSLine, RiArrowUpSLine, RiDeleteBinLine, RiExternalLinkLine, RiLockLine } from '@remixicon/vue'
 import AdminBadge from '~/components/AdminBadge/AdminBadge.vue'
 import DataserviceAccessTypeBadge from '~/components/AdminTable/AdminDataservicesTable/DataserviceAccessTypeBadge.vue'
@@ -283,6 +293,7 @@ import ContactPoint from '~/components/ContactPoint.vue'
 import OrganizationOwner from '~/components/OrganizationOwner.vue'
 import ReportModal from '~/components/Spam/ReportModal.vue'
 
+const config = useRuntimeConfig()
 const route = useRoute()
 const { formatDate } = useFormatDate()
 
@@ -301,6 +312,31 @@ const openSwagger = ref(false)
 const accessAudiences = computed(() => (['local_authority_and_administration', 'company_and_association', 'private'] as Array<DataserviceAccessAudienceType>)
   .map(type => dataservice.value.access_audiences.find(a => a.role === type))
   .filter(Boolean) as Array<DataserviceAccessAudience>)
+
+const metricsViews = ref<null | Record<string, number>>(null)
+const metricsViewsTotal = ref<null | number>(null)
+
+watchEffect(async () => {
+  if (!dataservice.value.id) return
+  const response = await fetch(`${config.public.metricsApi}/api/dataservices/data/?dataservice_id__exact=${dataservice.value.id}&metric_month__sort=desc&page_size=12`)
+  const page = await response.json()
+
+  const views: Record<string, number> = {}
+
+  for (const { metric_month, monthly_visit } of page.data) {
+    views[metric_month] = monthly_visit
+  }
+  // Fetching totals
+  const totalResponse = await fetch(`${config.public.metricsApi}/api/dataservices_total/data/?dataservice_id__exact=${dataservice.value.id}`)
+  const totalPage = await totalResponse.json()
+
+  let totalViews = 0
+  if (page.data[0]) {
+    totalViews = totalPage.data[0].visit
+  }
+  metricsViews.value = views
+  metricsViewsTotal.value = totalViews
+})
 
 onMounted(async () => {
   await redirectLegacyHashes([
